feat: use fixed-width integer types and bump Foundry to v0.1.0-alpha.25
Foundry v0.1.0-alpha.24 added uint, long, and ulong integer types, and
alpha.25 made int a 32-bit type. Update the protobuf codegen to map
protobuf integers by width and signedness instead of collapsing all ten
spellings to int:

  int32/sint32/sfixed32  -> int    (32-bit signed)
  uint32/fixed32         -> uint   (32-bit unsigned)
  int64/sint64/sfixed64  -> long   (64-bit signed)
  uint64/fixed64         -> ulong  (64-bit unsigned)

The runtime wire codec, read carriers, JSON serialization, and JSON
deserialization are all retyped to respect carrier boundaries (signed
INT and unsigned UINT cannot mix). Unsigned scalars get their own
varint/fixed codecs and carriers (VarintReadUnsigned, FixedReadUnsigned);
fixed64 and sfixed64 readers are split since their carriers no longer
meet. 32-bit fields narrow from the wide carrier with a checked `as`
cast.

CI Foundry release tag is bumped to v0.1.0-alpha.25 in both the
workflow and the install script default. The engine type table gains
uint, long, and ulong as builtins and records the alpha.25 source
version. Golden examples, JSON golden corpus, and well-known type
bindings are regenerated.

+++ examples/golden/cafecito/game/v1/PlayerStatus.pb.fs
@@ -11,7 +11,7 @@ enum_name PlayerStatus:
 		return self as int
 
 	## Returns the case for a protobuf wire value, or null if it names none.
-	static func from_wire(value: int) -> Self?:
+	static func from_wire(value: long) -> Self?:
 		match value:
 			0:
 				return PlayerStatus.PLAYER_STATUS_UNSPECIFIED
